Rename AppIntegration services to managers.
