+++ Moda.Services/Moda.AppIntegration/src/Moda.AppIntegration.Application/Interfaces/IAzureDevOpsBoardsImportManager.cs
@@ -1,7 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 
 namespace Moda.AppIntegration.Application.Interfaces;
-public interface IAzureDevOpsBoardsImportService : ITransientService
+public interface IAzureDevOpsBoardsImportManager : ITransientService
 {
     Task<Result> ImportOrganizationConfiguration(Guid connectionId, CancellationToken cancellationToken);
     Task<Result> InitWorkspaceIntegration(Guid connectionId, Guid workspaceExternalId, string workspaceKey, CancellationToken cancellationToken);
